[templates] extract DrawBoundaryUserAction to src/types
